[RHCLOUD-41916] cleanup drawer deprecated code
