+++ connector-drawer/src/main/java/com/redhat/cloud/notifications/connector/drawer/DrawerCloudEventDataExtractor.java
@@ -23,7 +23,6 @@ public class DrawerCloudEventDataExtractor extends CloudEventDataExtractor {
         exchange.setProperty(ExchangeProperty.UNSUBSCRIBERS, notification.unsubscribers());
         exchange.setProperty(ExchangeProperty.UNSUBSCRIBERS, notification.unsubscribers());
         exchange.setProperty(ExchangeProperty.AUTHORIZATION_CRITERIA, notification.authorizationCriteria());
-        exchange.setProperty(ExchangeProperty.USE_SIMPLIFIED_ROUTE, drawerConnectorConfig.useSimplifiedRoute(notification.orgId()));
         exchange.setProperty(ExchangeProperty.EVENT_DATA, notification.eventData());
     }
 }
